Activity logging, password security, search, social login
Plus a minor fix to the domain resolution logic.

+++ database/migrations/2018_08_18_201318_create_activity_log_table.php
@@ -11,12 +11,12 @@ class CreateActivityLogTable extends Migration
     public function up()
     {
         Schema::create(config('activitylog.table_name'), function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('log_name')->nullable();
             $table->text('description');
-            $table->integer('subject_id')->nullable();
+            $table->unsignedBigInteger('subject_id')->nullable();
             $table->string('subject_type')->nullable();
-            $table->integer('causer_id')->nullable();
+            $table->unsignedBigInteger('causer_id')->nullable();
             $table->string('causer_type')->nullable();
             $table->text('properties')->nullable();
             $table->timestamps();
