+++ database/migrations/2018_08_18_201318_create_tag_tables.php
@@ -8,17 +8,17 @@ class CreateTagTables extends Migration
     public function up()
     {
         Schema::create('tags', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->json('name');
             $table->json('slug');
             $table->string('type')->nullable();
-            $table->integer('order_column')->nullable();
+            $table->bigInteger('order_column')->nullable();
             $table->timestamps();
         });
 
         Schema::create('taggables', function (Blueprint $table) {
-            $table->integer('tag_id')->unsigned();
-            $table->integer('taggable_id')->unsigned();
+            $table->unsignedBigInteger('tag_id')->unsigned();
+            $table->unsignedBigInteger('taggable_id')->unsigned();
             $table->string('taggable_type');
 
             $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade');
